requirements.txt
fastapi
uvicorn
psycopg2
python-dotenv
PyPDF2
pytesseract
pdf2image
requests 